+++ src/llm_openai.jl
@@ -498,13 +498,18 @@ const OPENAI_TOKEN_IDS = Dict("true" => 837,
     "20" => 508)
 
 """
-    encode_choices(schema::OpenAISchema, choices::AbstractVector{<:Tuple{<:AbstractString, <:AbstractString}}; kwargs...)
+    encode_choices(schema::OpenAISchema, choices::AbstractVector{<:AbstractString}; kwargs...)
+
+    encode_choices(schema::OpenAISchema, choices::AbstractVector{T};
+    kwargs...) where {T <: Tuple{<:AbstractString, <:AbstractString}}
 
 Encode the choices into an enumerated list that can be interpolated into the prompt and creates the corresponding logit biases (to choose only from the selected tokens).
 
+Optionally, can be a vector tuples, where the first element is the choice and the second is the description.
+
 # Arguments
 - `schema::OpenAISchema`: The OpenAISchema object.
-- `choices::AbstractVector{<:Tuple{<:AbstractString, <:AbstractString}}`: The choices to be encoded, represented as a vector of tuples where each tuple contains a choice and its description.
+- `choices::AbstractVector{<:Union{AbstractString,Tuple{<:AbstractString, <:AbstractString}}}`: The choices to be encoded, represented as a vector of the choices directly, or tuples where each tuple contains a choice and its description.
 - `kwargs...`: Additional keyword arguments.
 
 # Returns
@@ -521,6 +526,13 @@ logit_bias # Output: Dict(837 => 100, 905 => 100)
 choices_prompt, logit_bias, _ = PT.encode_choices(PT.OpenAISchema(), ["animal", "plant"])
 choices_prompt # Output: "1. \"animal\"\n2. \"plant\""
 logit_bias # Output: Dict(16 => 100, 17 => 100)
+```
+
+Or choices with descriptions:
+```julia
+choices_prompt, logit_bias, _ = PT.encode_choices(PT.OpenAISchema(), [("A", "any animal or creature"), ("P", "for any plant or tree"), ("O", "for everything else")])
+choices_prompt # Output: "1. \"A\" for any animal or creature\n2. \"P\" for any plant or tree\n3. \"O\" for everything else"
+logit_bias # Output: Dict(16 => 100, 17 => 100, 18 => 100)
 ```
 """
 function encode_choices(schema::OpenAISchema,
@@ -541,8 +553,35 @@ function encode_choices(schema::OpenAISchema,
 
     return join(choices_prompt, "\n"), logit_bias, choices
 end
+function encode_choices(schema::OpenAISchema,
+        choices::AbstractVector{T};
+        kwargs...) where {T <: Tuple{<:AbstractString, <:AbstractString}}
+    global OPENAI_TOKEN_IDS
+    ## if all choices are in the dictionary, use the dictionary
+    if all(x -> haskey(OPENAI_TOKEN_IDS, x), choices)
+        choices_prompt = ["$c for \"$desc\"" for (c, desc) in choices]
+        logit_bias = Dict(OPENAI_TOKEN_IDS[c] => 100 for (c, desc) in choices)
+    elseif length(choices) <= 20
+        ## encode choices to IDs 1..20
+        choices_prompt = ["$(i). \"$c\" for $desc" for (i, (c, desc)) in enumerate(choices)]
+        logit_bias = Dict(OPENAI_TOKEN_IDS[string(i)] => 100 for i in 1:length(choices))
+    else
+        ArgumentError("The number of choices must be less than or equal to 20.")
+    end
 
-function decode_choices(schema::OpenAISchema, choices, conv::AbstractVector)
+    return join(choices_prompt, "\n"), logit_bias, first.(choices)
+end
+
+# For testing
+function encode_choices(schema::TestEchoOpenAISchema, args...; kwargs...)
+    return encode_choices(OpenAISchema(), args...; kwargs...)
+end
+# For testing
+function decode_choices(schema::TestEchoOpenAISchema, args...; kwargs...)
+    return decode_choices(OpenAISchema(), args...; kwargs...)
+end
+
+function decode_choices(schema::OpenAISchema, choices, conv::AbstractVector; kwargs...)
     if length(conv) > 0 && last(conv) isa AIMessage
         conv[end] = decode_choices(schema, choices, last(conv))
     end
@@ -552,7 +591,7 @@ end
 """
     decode_choices(schema::OpenAISchema,
         choices::AbstractVector{<:AbstractString},
-        msg::AIMessage)
+        msg::AIMessage; kwargs...)
 
 Decodes the underlying AIMessage against the original choices to lookup what the category name was.
 
@@ -560,7 +599,7 @@ If it fails, it will return `msg.content == nothing`
 """
 function decode_choices(schema::OpenAISchema,
         choices::AbstractVector{<:AbstractString},
-        msg::AIMessage)
+        msg::AIMessage; kwargs...)
     global OPENAI_TOKEN_IDS
     parsed_digit = tryparse(Int, strip(msg.content))
     if !isnothing(parsed_digit) && haskey(OPENAI_TOKEN_IDS, strip(msg.content))
@@ -587,33 +626,31 @@ Classifies the given prompt/statement into an arbitrary list of `choices`, which
 It's quick and easy option for "routing" and similar use cases, as it exploits the logit bias trick and outputs only 1 token.
 classify into an arbitrary list of categories (including with descriptions). It's quick and easy option for "routing" and similar use cases, as it exploits the logit bias trick, so it outputs only 1 token.
 
-Choices are rewritten into an enumerated list and mapped to a few known OpenAI tokens (maximum of 20 choices supported).
+Choices are rewritten into an enumerated list and mapped to a few known OpenAI tokens (maximum of 20 choices supported). Mapping of token IDs for GPT3.5/4 are saved in variable `OPENAI_TOKEN_IDS`.
 
-Note: Credit goes to [AAAzzam](https://twitter.com/AAAzzam/status/1669753721574633473) for the original logit bias trick.
-
-It uses Logit bias trick and limits the output to 1 token to force the model to output only true/false/unknown.
-
-Output tokens used (via `api_kwargs`):
-- 837: ' true'
-- 905: ' false'
-- 9987: ' unknown'
+It uses Logit bias trick and limits the output to 1 token to force the model to output only true/false/unknown. Credit for the idea goes to [AAAzzam](https://twitter.com/AAAzzam/status/1669753721574633473).
 
 # Arguments
 - `prompt_schema::AbstractOpenAISchema`: The schema for the prompt.
-- `prompt`: The prompt/statement to classify if it's a `String`. If it's a `Symbol`, it is expanded as a template via `render(schema,template)`.
+- `prompt`: The prompt/statement to classify if it's a `String`. If it's a `Symbol`, it is expanded as a template via `render(schema,template)`. Eg, templates `:JudgeIsItTrue` or `:InputClassifier`
+- `choices::AbstractVector{T}`: The choices to be classified into. It can be a vector of strings or a vector of tuples, where the first element is the choice and the second is the description.
 
 # Example
 
+Given a user input, pick one of the two provided categories:
 ```julia
 choices = ["animal", "plant"]
 input = "Palm tree"
-aiclassify(tpl; choices, input)
+aiclassify(:InputClassifier; choices, input)
+```
 
+Choices with descriptions provided as tuples:
+```julia
 choices = [("A", "any animal or creature"), ("P", "for any plant or tree"), ("O", "for everything else")]
 input = "spider"
 input = "daphodil"
 input = "castle"
-aiclassify(tpl; choices, input)
+aiclassify(:InputClassifier; choices, input)
 ```
 
 You can still use a simple true/false classification: